Merge dev to master

+++ server/controllers/user/logout.js
@@ -1,20 +1,19 @@
-const db = require("../../db");
-const jwt = require("jsonwebtoken");
-
+const db = require('../../db');
+const jwt = require('jsonwebtoken');
+const { getAccessToken } = require('../../funcs/index');
 module.exports = async (req, res) => {
   const accessToken = req.cookies.accessToken;
-  const verified = jwt.verify(
-    accessToken,
-    process.env.ACCESS_SECRET,
-    (err, decoded) => {
-      if (err) return null;
-      return decoded;
+  const verified = await getAccessToken(req, res);
+  if (verified) {
+    if (verified.isGuest) {
+      await db.query('DELETE FROM users WHERE id=?', [verified.id]);
     }
-  );
-
-  if (verified.isGuest) {
-    await db.query('DELETE FROM users WHERE id=?', [verified.id]);
+    res.clearCookie('accessToken', { path: '/' });
+    return res.status(204).end();
   }
-  res.clearCookie("accessToken", { path: "/" });
-  res.status(205).json({ message: "로그아웃되었습니다" });
+  return res.status(403).json({
+    data: null,
+    path: '/users/logout',
+    message: 'forbidden access',
+  });
 };
